+++ src/azure-cli-core/azure/cli/core/_profile.py
@@ -324,6 +324,20 @@ class Profile(object):
                 str(account[_SUBSCRIPTION_ID]),
                 str(account[_TENANT_ID]))
 
+    def get_refresh_token(self, resource=CLOUD.endpoints.active_directory_resource_id,
+                          subscription=None):
+        account = self.get_subscription(subscription)
+        user_type = account[_USER_ENTITY][_USER_TYPE]
+        username_or_sp_id = account[_USER_ENTITY][_USER_NAME]
+
+        if user_type == _USER:
+            _, _, token_entry = self._creds_cache.retrieve_token_for_user(
+                username_or_sp_id, account[_TENANT_ID], resource)
+            return None, token_entry[_REFRESH_TOKEN], str(account[_TENANT_ID])
+
+        sp_secret = self._creds_cache.retrieve_secret_of_service_principal(username_or_sp_id)
+        return username_or_sp_id, sp_secret, str(account[_TENANT_ID])
+
     def get_raw_token(self, resource=CLOUD.endpoints.active_directory_resource_id,
                       subscription=None):
         account = self.get_subscription(subscription)
@@ -371,21 +385,6 @@ class Profile(object):
         result['subscriptionId'] = result.pop('id')
         result['endpoints'] = CLOUD.endpoints
         return result
-
-    def get_refresh_credentials(self, resource=CLOUD.endpoints.management,
-                                subscription_id=None):
-        account = self.get_subscription(subscription_id)
-        user_type = account[_USER_ENTITY][_USER_TYPE]
-        username_or_sp_id = account[_USER_ENTITY][_USER_NAME]
-
-        if user_type == _USER:
-            refresh_object = self._creds_cache.retrieve_token_entry_for_user(
-                username_or_sp_id, account[_TENANT_ID], resource)[_REFRESH_TOKEN]
-        else:
-            refresh_object = self._creds_cache \
-                .retrieve_cred_for_service_principal(username_or_sp_id)
-
-        return refresh_object
 
     def get_installation_id(self):
         installation_id = self._storage.get(_INSTALLATION_ID)