+++ src/command_modules/azure-cli-acr/HISTORY.rst
@@ -2,6 +2,12 @@
 
 Release History
 ===============
+unreleased
+++++++++++
+* Add managed registries with Managed SKU.
+* Add webhooks for managed registries with acr webhook command module.
+* Add AAD authentication with acr login command.
+* Add delete command for docker repositories, manifests, and tags.
 
 2.0.6 (2017-06-13)
 ++++++++++++++++++